requirements.txt
Django>=4.2,<5.2
Pillow>=10.0.0
argon2-cffi>=21.3.0
razorpay>=1.4.0
reportlab>=4.0.0

# Background tasks & scheduling
celery>=5.3.0
celery[redis]>=5.3.0
django-celery-beat>=2.5.0

# Caching
redis>=5.0.0

# REST API
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0

# SMS (optional — only needed if TWILIO_ACCOUNT_SID is set)
twilio>=8.10.0
